BUG en Eliminar
No eliminaba casas

+++ admin/borrar.php
@@ -1,10 +1,18 @@
        <?php 
        include('conexion.php');
        $id = @$_GET['id'];
-
-        mysql_query("DELETE FROM noticias WHERE id = '$id'"); 
-        echo "Publicación Eliminada";
+       $tipo = @$_GET['tipo'];
+       if ($tipo=="noticia") {
+       	
+	        mysql_query("DELETE FROM noticias WHERE id = '$id'"); 
+	        echo "Publicación Eliminada";
+    	}else
+    	{
+    		 mysql_query("DELETE FROM inmuebles WHERE id = '$id'"); 
+        	echo "Inmueble Eliminado";
+    	}
         ?>
+
         <div class="back">
             
             <a href="buscar.php">Regresar</a>
